@@ -1,3 +1,4 @@
+/*
 using Microsoft.AspNetCore.Mvc;
 using System.Security.Claims;
 using SS.Backend.DataAccess;
@@ -15,7 +16,32 @@ public class SecurityController : Controller
         _authService = authService;
     }
     
+    [HttpPost("postAuthN")]
+    [Route("/authN")]
+    public async Task<IActionResult> Authenticate([FromBody] AuthenticationRequest request)
+        {
+            var (principal, response) = await _authService.Authenticate(request);
+
+            if (response.HasError)
+            {
+                return BadRequest(response.ErrorMessage);
+            }
+
+            return Ok(new { principal });
+        }
+
+    [HttpGet(Name = "getAuthZ")]
+    [Route("/authZ")]
+    public async Task<IActionResult> AuthZ([FromRouteAttribute] SSPrincipal currentPrincipal, IDictionary<string, string> requiredClaims)
+    {
+        var verify = await _authService.IsAuthorize(currentPrincipal, requiredClaims);
+
+        return Ok(new { verify} );
+    }
+
+    
     [HttpPost(Name = "postSendOTP")]
+    [Route("/otp")]
     public async Task<IActionResult> SendOTP([FromBody] AuthenticationRequest authRequest)
     {   
         var (otp, response) = await _authService.SendOTP_and_SaveToDB(authRequest);
@@ -27,53 +53,61 @@ public class SecurityController : Controller
         
         return Ok(new { otp });
     }
-
-    [HttpPost(Name = "postAuthN")]
-    public async Task<IActionResult> AuthN ([FromBody] AuthenticationRequest authRequest)
-    {
-        var (principal, response) = await _authService.Authenticate(authRequest);
-
-        if(response.HasError)
-        {
-            return BadRequest(response.ErrorMessage);
-        }
-
-        return Ok(new { principal });
-    }
-
-    [HttpPost(Name = "postAuthZ")]
-    public async Task<IActionResult> AuthZ([FromRouteAttribute] SSPrincipal currentPrincipal, IDictionary<string, string> requiredClaims)
-    {
-        var verify = await _authService.IsAuthorize(currentPrincipal, requiredClaims);
-
-        return Ok(new { verify} );
-    }
-
-    /* 
-    [HttpGet(Name = "getAuthN")]
-    public async Task<ActionResult<List<AuthenticationRequest>>> Authenticate(AuthenticationRequest authRequest)
-    {
-        var response = await _authService.Authenticate(authRequest);
-        
-        if (response.HasError)
-        {
-            Console.WriteLine(response.ErrorMessage);
-            return StatusCode(500, response.ErrorMessage);
-        }
-
-        var requests = response.ValuesRead.Select(row => new AuthenticationRequest
-        {
-            RequestId = (int)row[0],
-            UserHash = (string)row[1],
-            RequestDate = (DateTime)row[2],
-            Status = (string)row[3],
-            RequestType = (string)row[4],
-            ResolveDate = row[5] != DBNull.Value ? DateTime.Parse((string)row[5]) : (DateTime?)null,
-            AdditionalInformation = row[6] != DBNull.Value ? (string)row[6] : null
-        }).ToList();
-
-        return Ok(requests);
-    }
-    */
+    
     
 }
+*/
+
+using Microsoft.AspNetCore.Mvc;
+using SS.Backend.Security;
+
+namespace AuthAPI.Controllers
+{
+    [ApiController]
+    [Route("api/auth")]
+    public class AuthNController : Controller
+    {
+        private readonly SSAuthService _authService;
+
+        public AuthNController(SSAuthService authService)
+        {
+            _authService = authService;
+        }
+
+        [HttpPost("sendOTP")]
+        public async Task<IActionResult> SendOTP([FromBody] AuthenticationRequest request)
+        {
+            var (otp, response) = await _authService.SendOTP_and_SaveToDB(request);
+
+            if (response.HasError)
+            {
+                return BadRequest(response.ErrorMessage);
+            }
+
+            return Ok(new { otp });
+        }
+
+        [HttpPost("authenticate")]
+        public async Task<IActionResult> Authenticate([FromBody] AuthenticationRequest request)
+        {
+            var (principal, response) = await _authService.Authenticate(request);
+
+            if (response.HasError)
+            {
+                return BadRequest(response.ErrorMessage);
+            }
+
+            return Ok(new { principal });
+        }
+
+        [HttpGet(Name = "getAuthZ")]
+        [Route("/authZ")]
+        public async Task<IActionResult> AuthZ([FromRouteAttribute] SSPrincipal currentPrincipal, IDictionary<string, string> requiredClaims)
+        {
+            var verify = await _authService.IsAuthorize(currentPrincipal, requiredClaims);
+
+            return Ok(new { verify} );
+        }
+    }
+}
+
